feat: update version to 2.18.54, refactor storage imports to use engine directory, and remove legacy FileManager and FolderManager
