+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "axiodb",
-  "version": "2.17.54",
+  "version": "2.18.54",
   "description": "A blazing-fast, lightweight, and scalable nodejs package based DBMS for modern application. Supports schemas, encryption, and advanced query capabilities.",
   "main": "./lib/config/DB.js",
   "types": "./lib/config/DB.d.ts",
