+++ source/Operation/CRUD Operation/Delete.operation.ts
@@ -7,8 +7,8 @@ import {
 import Converter from "../../Helper/Converter.helper";
 import { CryptoHelper } from "../../Helper/Crypto.helper";
 import ResponseHelper from "../../Helper/response.helper";
-import FileManager from "../../Storage/FileManager";
-import FolderManager from "../../Storage/FolderManager";
+import FileManager from "../../engine/Storage/FileManager";
+import FolderManager from "../../engine/Storage/FolderManager";
 
 // Import All Utility
 import HashmapSearch from "../../utils/HashMapSearch.utils";
