+++ source/Operation/CRUD Operation/Update.operation.ts
@@ -8,8 +8,8 @@ import Converter from "../../Helper/Converter.helper";
 import { CryptoHelper } from "../../Helper/Crypto.helper";
 import ResponseHelper from "../../Helper/response.helper";
 import { SchemaTypes } from "../../Models/DataTypes.models";
-import FileManager from "../../Storage/FileManager";
-import FolderManager from "../../Storage/FolderManager";
+import FileManager from "../../engine/Storage/FileManager";
+import FolderManager from "../../engine/Storage/FolderManager";
 import HashmapSearch from "../../utils/HashMapSearch.utils";
 import Sorting from "../../utils/SortData.utils";
 import { Console } from "outers";
